server/pyproject.toml
[tool.poetry]
name = "testsystemserver"
version = "0.1.0"
description = ""
authors = ["ysyyhhh <757578166@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.7.16"
cython = "3.0.3"
asgiref = "3.7.2"
attrs = "23.1.0"
backports-cached-property = "1.0.2"
build = "0.10.0"
cachecontrol = "0.12.14"
certifi = "2023.7.22"
charset-normalizer = "3.3.0"
cleo = "2.0.1"
colorama = "0.4.6"
crashtest = "0.4.1"
cycler = "0.11.0"
distlib = "0.3.7"
django = "3.2.22"
django-cors-headers = "4.1.0"
docopt = "0.6.2"
dulwich = "0.21.6"
filelock = "3.12.2"
fonttools = "4.38.0"
fsspec = "2023.1.0"
html5lib = "1.1"
huggingface-hub = "0.16.4"
idna = "3.4"
importlib-metadata = "6.7.0"
importlib-resources = "5.12.0"
installer = "0.7.0"
jaraco-classes = "3.2.3"
jsonschema = "4.17.3"
keyring = "23.13.1"
kiwisolver = "1.4.5"
lockfile = "0.12.2"
matplotlib = "3.5.3"
more-itertools = "9.1.0"
msgpack = "1.0.5"
numpy = "1.21.6"
opencv-python = "4.8.1.78"
packaging = "23.2"
pexpect = "4.8.0"
pillow = "9.5.0"
pipreqs = "0.4.13"
pkginfo = "1.9.6"
pkgutil-resolve-name = "1.3.10"
platformdirs = "3.11.0"
poetry = "1.5.1"
poetry-core = "1.6.1"
poetry-plugin-export = "1.4.0"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pyparsing = "3.1.1"
pyproject-hooks = "1.0.0"
pyrsistent = "0.19.3"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
pywin32-ctypes = "0.2.2"
pyyaml = "6.0.1"
rapidfuzz = "2.15.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
safetensors = "0.4.0"
shapely = "2.0.1"
shellingham = "1.5.3"
six = "1.16.0"
sqlparse = "0.4.4"
timm = "0.9.7"
tomli = "2.0.1"
tomlkit = "0.12.1"
torch = "1.13.1"
torchvision = "0.14.1"
tqdm = "4.66.1"
trove-classifiers = "2023.9.19"
typing-extensions = "4.7.1"
urllib3 = "1.26.17"
virtualenv = "20.24.5"
webencodings = "0.5.1"
yarg = "0.1.9"
zipp = "3.15.0"
open3d = "^0.17.0"
scipy = "1.7.3"
wand = "0.6.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
